requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
moviepy
opencv-python-headless
numpy
scikit-learn
requests
httpx
python-telegram-bot
pytest